Un-promisify everything (except .svg files, which are broken)
Because we are now doing the file transforms in the Pug filter, we have
to be syncronous. All the filter-specific transforms have been moved to
the transform() method.

Note that .json and .yml files are NOT in the transform() method, they
stay as Pug variables.

See deasync for a possible (ugly) workaround for svgo:

https://github.com/svg/svgo/issues/584#issuecomment-265965377