add average as an aggregation method
The default is still "sum", but if you add "avg" to the "groupBy" text
input, that will be used instead.

I briefly had this as a radio group in the web ui, but it really only
matters if a "groupBy" exists, so it was more natural to just include it
in that value.